temporal-tutorial/py_client/requirements.txt
# Temporal SDK
temporalio==1.14.1

# Web Framework for UI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
jinja2==3.1.3

# OpenTelemetry SDK and Instrumentation
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1

# Additional dependencies for production
python-dotenv==1.0.0